requirements.txt
requests
kafka-python
pandas
apache-airflow